DDF-1897 - Moved Versioning into its own module
