+++ catalog/core/catalog-core-versioning/versioning-common/src/test/groovy/ddf/catalog/core/versioning/HistoryMetacardImplSpecTest.groovy
@@ -1,0 +1,5 @@
+package ddf.catalog.core.versioning
+
+
+class HistoryMetacardImplSpecTest extends spock.lang.Specification {
+}
